requirements.txt
anyio==3.3.2
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
crypto==1.4.1
cryptography==35.0.0
Django==3.2.8
django-environ==0.7.0
django-rest-knox==4.1.0
django-versatileimagefield==2.2
djangorestframework==3.12.4
eth-hash==0.3.2
eth-typing==2.2.2
Flutterwave==1.0.7
fpdf==1.7.2
geojson==2.5.0
h11==0.12.0
idna==3.2
multidict==4.7.6
Naked==0.1.31
Pillow==8.3.2
pycparser==2.20
pycrypto==2.6.1
pydantic==1.8.2
PySocks==1.7.1
python-magic-bin==0.4.14
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
simplejson==3.17.5
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.2.4
starlette==0.16.0
stripe==2.60.0
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==8.1